💫 UPDATE: move tooltip from count component to ReaderUnreadCount
