+++ client/layout/sidebar/reader-unread-count.tsx
@@ -1,5 +1,7 @@
-import { Count } from '@automattic/components';
+import { Count, Tooltip } from '@automattic/components';
 import { useTranslate } from 'i18n-calypso';
+import { useState } from 'react';
+import type { MouseEvent } from 'react';
 
 interface ReaderUnreadCountProps {
 	count?: number;
@@ -7,6 +9,7 @@ interface ReaderUnreadCountProps {
 
 const ReaderUnreadCount = ( { count }: ReaderUnreadCountProps ): JSX.Element | null => {
 	const translate = useTranslate();
+	const [ anchor, setAnchor ] = useState< HTMLSpanElement | null >( null );
 
 	if ( ! count || count < 1 ) {
 		return null;
@@ -18,7 +21,23 @@ const ReaderUnreadCount = ( { count }: ReaderUnreadCountProps ): JSX.Element | n
 		textOnly: true,
 	} );
 
-	return <Count count={ count } tooltipText={ tooltipText } aria-label={ tooltipText } compact />;
+	return (
+		<>
+			<Count
+				count={ count }
+				aria-label={ tooltipText }
+				compact
+				onMouseEnter={ ( event: MouseEvent< HTMLSpanElement > ) =>
+					setAnchor( event.currentTarget )
+				}
+				onMouseLeave={ () => setAnchor( null ) }
+			/>
+
+			<Tooltip context={ anchor } focusOnShow={ false } isVisible={ !! anchor } showDelay={ 500 }>
+				{ tooltipText }
+			</Tooltip>
+		</>
+	);
 };
 
 export default ReaderUnreadCount;
